midterm/ecommerce/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.2
django-celery-results==2.5.1
django-debug-toolbar==4.4.6
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kombu==5.4.2
pika==1.3.2
prompt_toolkit==3.0.48
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.1.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13